+++ lib/src/io/interface.dart
@@ -69,5 +69,8 @@ void ensureDir(String path) => null;
 /// [path].
 Iterable<String> listDir(String path) => null;
 
+/// Returns the modification time of the file at [path].
+DateTime modificationTime(String path) => null;
+
 /// Gets and sets the exit code that the process will use when it exits.
 int exitCode;